Remove unused code and move single case to related files
